legal-backend/requirements.txt
fastapi
uvicorn
python-dotenv
pandas
httpx
scikit-learn
pydantic
pydantic-settings